+++ backend/src/routes/driver.ts
@@ -7,7 +7,7 @@ import validateDriver from "../middlewares/validateDriver";
 
 const router = express(); 
 
-router.post("/add-driver", validateDriver,  async (req : Request, res : Response) => {
+router.post("/add", validateDriver,  async (req : Request, res : Response) => {
     try{ 
         const { 
             email, 
@@ -54,7 +54,6 @@ router.post("/add-driver", validateDriver,  async (req : Request, res : Response
     res.status(200).json({
         title : "Driver Added!", 
         message : `Driver ${last_name}, ${first_name} ${middle_name} has been added`,
-
         
     })
 
@@ -69,12 +68,46 @@ router.post("/add-driver", validateDriver,  async (req : Request, res : Response
         console.error("Unexpected error occurred:", error);
         res.status(500).json({ title: "Server Error", message: "An unexpected error occurred." });
         }
-        
     }
 
-    
-  
-
 });
+
+router.get("/get", async (req: Request, res: Response) => {
+
+    try {
+      console.log("Fetching drivers from the database...");
+
+      const { rows: drivers } = await pool.query(
+       ` SELECT first_name, 
+        last_name,
+        middle_name, 
+        email, 
+        sex, 
+        driver_type, 
+        license_number, 
+        license_expiration_date 
+        FROM drivers`
+      );
+      console.log("Drivers fetched successfully:", drivers);
+  
+      // Send the drivers list as a response
+      res.json(drivers);
+
+    } catch (error) {
+      const errorMessage = (error as Error).message;
+      console.error("Error fetching drivers:", errorMessage); // Log the error for debugging
+      res.status(500).json({ title: "Unknown Error", message: errorMessage });
+    }
+  });
+
+  router.delete("/delete", async (req : Request, res : Response) => {
+
+    try{ 
+        console.log("Fetching. . .");
+    }
+    catch(error){ 
+        console.log("Deleted!")
+    }
+  })
 
 export default router;
