+++ frontend/src/hooks/useDrivers.ts
@@ -43,7 +43,7 @@ const useDrivers = () => {
   const fetchDrivers = async () => {
     setLoading(true);
     try {
-      const response = await fetch(`http://localhost:4444/api/driverslist`, {
+      const response = await fetch(`http://localhost:4444/driver/get`, {
         method: "GET", // Use GET for fetching data
         headers: {
           "Content-Type": "application/json",
